Support more weird names on Windows
Most filename restricitons are somewhere at the API layer, not at
the filesystem layer, which means that we can run into invalid
filenames during a scan.

It is actually happening for a bunch of users who have filenames
with trailing spaces and periods.

Unfortunately, as explained at length in InjectableFile the current
situation is not perfect. Ideally we'd support (allthethings) that
the underlying FS does. Turns out we can't with a vanilla JDK so
we'll either have to patch the JDK or add more methods to the driver.

In any case, this fix the immediate problem. The more exotic names
will be ignored, which is admittedly suboptimal but safe.

Change-Id: I9be2c1d44303a24d1dcc7bdc136e83136417810d