Reduced shared folder dialog flickering when updating role/deleting member.
Long story. Originally, we'd thought that flickering is caused by grossly
inefficiency in how we refresh the table after an update/delete.

The original impl. was to modify the underlying collection and then refresh
the whole table. The new impl. is to modify the underlying collection, and
only refresh the affected item.

I refactored how members are organized to make it easier to do this change,
only to learn that the true cause of the flickering was something else.
Since the refactoring is for the better and already done, I'm commiting it.

So the root cause of flickering was that when a round trip to SP is fast,
we'll disable and enable the table in a short period of time (~100ms).
The visual effect is that the table appears to flicker.

There's not much we can do about it in this case, short of adding
artificial delay. In practice, it's unlikely for the round trip to SP to be
fast enough to cause an issue, unless you happen to be on local-prod where
this is first observed.

Change-Id: I669a134f96d17f13fe55c11cfb44dba6d34d6531