Fix for weird enterprise-greeter bug
The observed behavior is fragile and subtle; here we go.

If a private cloud appliance is started in an environment with no
DHCP server, and then subsequently reconfigured with static IP
information, resolvconf becomes a loaded gun. It is easy
to tip the machine over into a broken state.

Certain actions cause the resolver to fire. When resolvconf wakes up,
he will put his cached resolv.conf into /etc/resolv.conf, totally
ignoring the values in /etc/network/ and breaking the appliance.

I believe the action has to do with hostname resolution failure but
that isn't proven. Examples that caused resolv.conf to be updated
included restarting the bootstrap service and restarting mysql.
The db-restore task is not involved, except that the bootstrap
task running without engaging maintenance mode caused the failure
with 100% certainty.

It appears a simple fix is to disable resolvconf updates when in
static mode. This commit does that; and re-enables those updates
when setting DHCP mode. This was tested and found efficacious.

A workaround for affected systems is to reboot the machine.

Change-Id: Ica56f107a08b8d623882c14cbf4e61ad1adcacd6