ACL: discrete roles -> independent flags
This commit changes the internal model but
leaves the UX largely unchanged.

Also refactor sharing rules to pave the way
to the brave new world of group-based and
DSL-defined customer-specific sharing rules.

The most significant change there is the new
warning/error mechanism.

For now simply update the BB-specific sharing
rules to be saner and closer to their actual
requirements.

This big internal change is a good opportunity
to bump SP protocol number and cleanup some
of the legacy cruft. If all goes according to
plan the next backward-incompatible change of
the SP protocol will be a switch to a RESTful
API.

Change-Id: I5534b1bd3d52f84ebd993d808919d3730d86ff0e