Add a "sign up all invited" command to signup.sh
The "-a" option will sign up all invited users in one fell swoop,
with the same password (see the -p option).

This is useful for signing up multiple users in a private cloud
appliance that doesn't allow self-signup. And it means you can sign
up users with non-deliverable email addresses.

Change-Id: I88108ca8dd34a181a5a645c4671cff87b8d1777d