Deleted unreliable unit tests.
Long story short, the unit test exposed a flaw in our logic on a feature we
don't care enough about. Since we can't get the unit test to reliably work
nor do we want to invest in this any more than to just get the unit test
to pass, we've decided the sane choice is to simply get rid of the unit
test.

Change-Id: Ib5874ec58283d47614b658d472a629eaec11910d