Fix deletion of conflict and NROs
When a folder is deleted from the filesystem,
the MAP phy op is used, to indicate that no
actual deletion needs to be performed but instead
data structures should be updated in reaction to
a deletion.

This is all fine and dandy for representable files
and folders but it breaks down for conflict branches
and NROs, resulting in ghost physical objects that
can never be cleaned up.

Change-Id: Iab57c4a11f6ded88c48448df2e69f6e4606135ce