Fix NRO and aliasing/migration interaction
Not terribly happy with the fix as it is. More work will be
required to cleanup the physical storage interface. The
semantics of PhysicalOp are particularly problematic and
need a serious rethinking.

Add a DPUT to fix NRO db for users that were unfortunate
enough to migrate or alias NROs.

Change-Id: Id56b6a9e6c6acea6b9d87cd32ac9e2bc3cedd961