Rework splunk channel implementation
This uses the netty-recommended way to build an automatically-reconnecting
client channel. On channel close, we queue a reconnect attempt (on a delay).
The new connection comes wrapped in a new Channel instance, so we also
provide a channel-connected notifier to save a handle to the Channel
for subsequent sends.

This should be an overall reduction in implementation complexity.

Also increased the test coverage through the use of judicious
package-private entry points.

Change-Id: I1b07cf7a824601948d23e12caafa54b393d20a17