Config server: reject current license if expired
The design doc says that we should 400 in this case, but we currently
200.  This causes the web UI accept an expired license file that it
should reject.

The only flow this should affect is users trying to reconfigure
instances with expired licenses, in which case we'd rather fail fast
than fail during bootstrap.

A consideration that we should address in the future is "what
maintenance operations should be permitted by a user authenticating with
the shasum of the current but expired license"

Fixes ENG-305.

Change-Id: I7532e1bbfb3b0db3caf422764dc3832fce0875e3