Fixed shared folder members count not shown on Mavericks.
On OS X Mavericks, the shared folder dialog doesn't show the number of
members in a shared folder.

The cause is that the tab is not refreshed after the tab text is updated.
Unfortunately, the only way I've found to force a refresh is by resizing
the tab folder either manually or programmatically.

The SWT source code shows that SWT just redirects the call to GTK, so it
appears to be an issue in GTK's interactions with OS X.

Change-Id: I80e01b8bcae0437c8df2c8297b512249a44d7820