Prefer NFC to NFD on OSX
OSX filesystems use NFD. Because they are not normalization-
preserving we cannot use the same "contextual NRO" logic used
for case-sensitivity.

Most other OSes are normalization-preserving and default to NFC.
Therefore, for maximum interoperability between heterogenous
devices it is best to pretend that the OSX filesystem only uses
NFC.

This is what the JRE is doing and what we did before rolling out
NRO changes. The decision to pick NFD was well-intentioned (stay
as close to the actual physical storage as possible) but turned
out to be shortsigthed:
not only did it cause suboptimal UX among heterogenous devices,
it also caused a bad regression where all NFC files created before
the NRO rollout stopped syncing (see SUPPORT-1160).

This change in behavior requires a migration to avoid the same
kind of weirdness that resulted form the previous change.
Fortunately most of the work can be done within the scanner and
only new remotely created NFC files need special treatment.

Change-Id: Idf28fa30f05d85b2983758c88100a3e1c0c7133e