Fix RestService threading
Jersey logic can safely be offloaded to a thread pool, however
the handling of chunked HTTP requests need to remain in a Netty
I/O thread or madness will ensue (out of order chunks and other
niceties).

Bonuses

 * Defer sending 100 Continue to first attempt to read from
the request body. This can help save bandwidth when trying to
upload a file to a non-existing resource or with an expired
OAuth token.

 * Cleanup setup code of some restless services.

Change-Id: Iab416efc248263342db97579af21c60c201c399b