safetynet rollback: disable db tampering detection
When we roll safetynet back, we change the inode number of the DB.
This is dangerous in general, but safe for safetynet.

In fact, if we don't disable the db-corruption protection, then the GUI
will hang when the daemon suicides on launch until we click the "Force
Launch" button.  So this should make rollbacks smoother.

Change-Id: Ib3be265b0ef89c19e7e0a686e6d4629c2e27bc02