lizard: extract site views into blueprint
We're eventually going to need to have internal views (for AeroFS folks
to download license requests bundles and upload them and do other such
tasks) and external views (for customers to request licenses and
download them, but only for their organization).

For security reasons, I want to run the internal site as a completely
separate process on a separate port that's only accessible to the VPN,
but I'd really like to be able to share the SQLAlchemy models and other
pieces of code where feasible.

Blueprints are the idiomatic way to do this with Flask.

Change-Id: I34f50656cb65e468289d518df9ad6b3b076e09ae