lizard: include shim of admin site
The admin site runs on a separate port and is activated by passing
"internal" to the command line as argv[1].

We add a second upstart job to supervise a process running the admin
site, and the appropriate puppetry to make sure it gets launched and
relaunched.

We also add some really simple forms to test out the admin site.  Most
of these will be replaced in short order; this just tests out the basic
ability of the second backend to render things.

Change-Id: Iad00eb781023074bcb210b97f3a20281558d802c