Detect and handle downstream errors synchronously
An execution handler that maps the incoming work (on the netty IO thread)
to an outbound execution thread pool. The EventResource implementation is
updated to block on the netty response for each send.

Some other minor changes; reduced verbosity on a few noisy and meaningless
exceptions in audit path. Fixed the formatting on device-id audit events.

A simple unit test is added to check that an error from the Downstream
channelfuture will result in an error back to the client.

Change-Id: I3ea04520553b121270f09472e3b826041705f5a6