Revert "logrotate: check log hourly, rotate if size > 500M"
Unfortunately it does not work as expected. It is unclear from documentation but size option completely overrides any period option (daily/weekly/etc) so log only get rotated based on size condition.
Unfortunately there is no config parameter to rotate based on one of the two conditions: daily OR if size reaches limit.
One of the option here is to use separate size-based configuration for hourly rotation but it will require to create separate script in /etc/cron.hourly instead of symlinking
Anyway it will be separate commit if we decided to go this way.
Reverting this commit to restore daily rotation...

This reverts commit 06316f57ba9bd29cf43713c15025c25a147c35de.

Change-Id: I6f5e94fd403546e67fdd2abd8098bec1e2a3be11