Fix message fragmentation in TunnelHandler
Max payload size must take length field into account or the frame decoder
throws a fit, causing the connection to be closed halfway through writing
the message.

This fixes ENG-484. The reason this particular bug never hit us when
transfering large files is that file data is pre-chunked in smaller
packets, as opposed to JSON data which is currently written in one
go.

Add unit test for payload fragmentation and improve the auxiliary
test class in the process.

Change-Id: Icf3aa728a5b91bd2849373f8fe72988f0c4fdc01