update_cloudfront: guess mime type, work anywhere
This script would delete things from S3 if you ran it from a folder
other than ~/repos/aerofs/src/web.  A tad dangerous.

Also, when uploading CSS files, it would not specify the "text/css"
mimetype, so when the browser loaded the file, it would refuse to
interpret the CSS, since cloudfront served it up as "text/plain".
So we add --guess-mime-type to hopefully avoid this in the future.

Change-Id: If45216898abacd346421c8234d3af3911fb007bc