Fix overly aggressive optimization
User.setOrganization should ALWAYS cause a bump in the ACL epoch
of the Team Server user, even when the organization does not
change, otherwise the TS in private cloud will not auto-join the
root store of users created after it was installed.

Simple DB migration to automatically bump the ACL epoch of the
PRIVATE_ORG TS user. TS will be restarted when updating and
pick up the new epoch and any store they have missed.

Update unit tests to catch this bug.

Change-Id: Iedc6c31f3517e350aff2cec7a18c58fa4d6edca4