config: handle unicode fields from license file
license_file actually has always provided us Unicode strings, but until
we actually tested this with a unicode company name in a license file,
they didn't behave correctly.

Fortunately, Flask and Jinja are unicode-native, so we can just make
this particular string a unicode string, and then all is well.

Change-Id: Ibd98faf425b9fe176006c59663e43c84b2f9cbff