config: be robust to mangled external.properties
In particular, don't crash on:

1) Empty lines
2) Lines beginning with a #

As a bonus, when running main.py, run in debug mode.  This can help make
debugging issues easier.

As an addon bonus, make a couple of things relative paths to make it
easier to run on a dev's machine without having to install in a Linux
VM.  (You may still have to patch up the SSL cert and license file
locations.)

Change-Id: I25fb51bc310d1680df02154c30b53b38f6e0d294