lizard-admin: tool to release new appliance image
An admin can now change the release version with the click of a button.
Lizard will ensure that the ova and qcow2 are both present in the S3
bucket before allowing the new version to go live.

Lizard will not check that the new version dominates the old, in case we
ever need to roll back a release due to unforseen breakage.

Change-Id: Iecdcd8dd6c97ba3b59af87a2ab0a22528852d7be