Add unit tests for auditor downstream SSL
Unit tests for sending downstream data over tcp-ssl.

A small refactoring: added a class in libtest that allows creation of ephemeral
server SSL socket factories; the certificate used by the server is accessible
as a string, which we can use in any place we explicitly specify a server cert.

Used today in LDAP and Auditor testing.

Change-Id: Ie65a6d61ecfb1078053cb53600d300ed562f03ae