Grub: different grub config for different image formats.
This allows us to boot nicely into:

    1. OpenStack and ec2 when using RAW or QCOW2, and
    2. VBox and VMWare when using OVA or VDI.

Also note this means the OVA will _not_ work with cloud services and the
RAW image will _not_ work with VBox and VMWare.

The complicating issue here is that due to the fact that OpenStack can take 30
sec or more to boot, and we don't want our VMWare/VBox boot to have to wait on a
timeout every time we start up.

Change-Id: If0321268e6b9211ccafe71b03273c4222b561f13