lizard: prefer autofocus attribute to JS focusing
JS focus will steal focus from the user if they're faster to start
filling out the form than the page is to load, since the focus event
only fires on document.ready.

OTOH, HTML5's autofocus attribute works as soon as the DOM is created,
so there's no gap and no bad user experience.

Also, use autofocus sparingly.  The general guideline is to use it only
on pages where the entire purpose of the page is to complete a form.

Change-Id: I553858deecccef0e1c6774ed9f43035f61073541