Fix more TS ACL epoch issues
Whenever a new user is created, the ACL epoch of the TS user
of the relevant org MUST be bumped and a vk notif should be
sent to ensure that any and all TS already installed can auto-
join the root store of said user.

The previous commit fixed this problkem when the user signing
up had an outstanding org invite but missed the following cases:
 * user joining without org invite
 * user auto-provisioned through OpenID
 * user auto-provisioned through LDAP

Update unit tests to catch regressions for all these cases.

Change-Id: I683de231fc9765588174df19d7f10f4a3e979d95