This is Sparta! (picture Leonidas kicking SP in the pit)
Sparta is a REStful endpoint that offers a public API to the
SP database (users, devices, shared folders).

This commit lays down the foundation for future development
of Sparta:
 * module
 * build
 * packaging
 * deployment

The plan is to reuse all the existing business logic that
lives in libservlet.

The SP database will be co-owned, i.e. whichever server comes
up first will handle any required the migration. At this time
the migration tasks still live in the sp module and are copied
into sparta at packaging time.

Change-Id: I417029ec129b3c5a9f3905706b4b36a263bb54cb