web/bunker: don't require /etc/hosts nonsense
/etc/hosts is not an appropriate way to insert an abstraction layer.
Let's do that properly in the application code.

Wow, there are a lot of remote services that bunker and web hit!
 - config server
 - email/ldap verification server
 - oauth server
 - sanity (service status) server
 - bootstrap server

All of these get their own parameter in the appropriate *.ini files,
rather than having the URLs hardcoded to localhost in the source,
forcing us to attempt to abstract that by redefining localhost.

Tested against local prod, and then tested running web and bunker
locally against that appliance image.  All was well.

Change-Id: I08dda37aa87955497801edb12138a2995b1e2dcc