lizard: encode all emails as UTF-8
If a user or org has a non-latin1-clean name, then we'll trip up on the
encoding when trying to send them an email.

Additionally, we should prefer to not import a license if we cannot send
an email to the user about it.  You can choose between potentially
sending duplicate messages and potentially dropping messages, and I
think for notification emails, we'd much rather send dupes.

Change-Id: Ie3e464fdc975225a305694e9c4e5c350037ad537