Fixed an issue with web authentication.
SP and the web client both maintain separate sessions. SP maintains session
in memory whereas the web client maintains session using cookies. The web
client makes requests to SP on the users behalf, thus SP's session ID
is stored in the web session data aka cookie.

On every request to the web client, the web client will first extract
SP's session ID and validate the SP session. If the SP session is invalid
or expired, the web client throws an exception and redirects to login. This
is undesirable and makes no sense on some pages (say login).

To solve this, the web client now assumes the web user has no permissions
when the SP session is invalid. If certain user permission is required,
the web client will redirect users to login. Otherwise, the requests will
be processed as normal because no additional permission is required.

Change-Id: I7865200ca360fb893d59cd89878b5cdbadc2229a