Minor fix to the configuration loader
Previously, getStaticConfigInputStream() would return null if the config
file wasn't found, instead of throwing an exception as the caller expected.

This didn't have any real impact in the client behavior because the NPE
that would then be thrown would still be caught at an upper level and an
error message about config loading failure would be displayed to the user,
but it did make debugging a little harder when looking at stack traces
because the NPE was hiding the real cause of error.

I believe this commit makes the code slightly better, without changing any
behavior.

Change-Id: I99dbfe0cd7b6941ab27ab30c6c2b02231a802d75