Fix Collector bug
When the Collector runs out of CLIENT tokens to fetch data
from a remote peer it uses a token  reclamation listener to
resume collection as soon as possible. If the lucky listener
somehow does not acquire a token then other waiting listeners
would NOT be notified of the availability of a token.

When the first 5 (quota of CLIENT tokens) listeners ended up
not acquiring a token collection would stop until the daemon
was restarted.

Change-Id: Id16a4d1da18c82008bf0da1880bd3ee8044ea265