Improve logical/physical fs race condition check
Avoid check altogether for files that are in the root anchor
(conflict branches and NROs). They are not expected to change
because users should not mess with them on pain of death but
being overzealous can create persistent failure states that
can only be left by unlinking...

Better logging when race condition is detected.

Change-Id: I9c7e62bbe4b198447a95ff317c02ae76ba15fcab