UI: recertify expired devices
This corrects the old ULTRecertify task to create a UI element
that will handle all the user authentication types:
 - client / team server
 - ldap / local credential / openid
 - cli / gui

In fact, those multiply out, so there are 12 paths through this...

Some light refactoring was done to lean this all on the existing
SetupModel and SignInActor mechanisms. Tested exhaustively by hand.

Change-Id: Ibcc8be236b54d49e0e3cb1485491f859d19c4336