Increase LinkedStorage robustness
In the face of timestamp inconsistency between db and filesystem,
increase logging, and schedule rescans.

This is to ensure that even a bug which violates some of our
assumptions does not result in the daemon being stuck in a
livelock-like situation.

There is one known root cause for the symptoms addressed by
this change and unfortunately it does not look easy to work
around (let alone to solve) as it is a combination of stupid
behaviors distributed in FSEvents and 3rd-party apps: if an
app keeps a file open for a long time, FSEvents only emits
at most two notifications for that entire time, regardless
of the actual write pattern.

Change-Id: I21c3aacece8a8c59cf5d589ab260307781b647a8