Unbreak osx updater script
The race condition has been present for a long time, but it got really bad with
the new JRE. Presumably because the shutdown timing has changed somewhat?

Regardless of root cause, we have observed osx clients failing to succeed at
updates because the updater.sh script dies after killing the client. This
resolves that by avoiding use of builtin kill, whose return code handling is
not what you expect.

This does not fix the clients that already have this updater script.

Change-Id: I7b64fa55d4720f964aea663908bd13fb9b0e6742