Bifrost: generate admin access tokens on request
Add the 'organization.admin' scope. This scope will be required for
user management actions (and others that previously claimed they needed
'admin access').

This scope can only be granted if the caller is an admin.

API doc updates are included to match. Also some refactoring to enable the above.

AMENDED: test and comment improvements

Change-Id: I59d874623cf5d034b2f2afe705e0002e909d0891