deploy: use the artifact's build version
Otherwise, in the case of two people deploying HC and PC concurrently,
you might wind up deploying a version to the wrong place.  In
particular:

version counter is 0.8.21
HC packages 0.8.22
PC bumps version number to 0.8.22
HC deploys to 0.8.23 (but all the packages contain version 0.8.22)
???

This way, we only read the version from a remote location once at
package-time, and keep that version number until we bless it by
deploying it publicly.

Change-Id: I67bae843098aa38c1230bd5e66f44c01e5ccc446