Fix regression in FID adjustment
PhysicalOp, the gift that keeps on giving...

As already mentioned in a few comments in the code, our use of
PhysicalOp in aliasing and migration is smelly and fragile. We
will need to give some  serious thought to this interface at
some point in the future. For now this change ensures that the
FID  consistency invariants aren't violated by the use of NOP
deletion in sharing/migration.

Change-Id: I5c687e2a331f16a9f5c03d941647d692aef3badd