Auto-create anchors on Team Server
When the list of members in a store changes, TS will now automatically
create anchors in the appropriate root stores as needed. This ensures
that joined shared folders become available through the API without
having to wait for the anchor to be created on a regular desktop
client and make its way to the Team Server.

Change-Id: I1830c4ff5d9605320a503067a73f596502ec360f