More robust handling of deletion for ignored objects
We recently updated the IgnoreList and didn't clean up the DB afterwards
This leaves some inconsistencies that LinkedStorage was not equipped to
deal with which in turn caused crash loops for some unlucky users.

NB: API users can currently create ignored files and folders, however
the corresponding logical objects will not survive for very long as
the scanner iwll add them to the TimeoutDeletionBuffer which will delete
them after 6s. However the corresponding physical objects will remain
until a non-ignored parent is deleted.

Change-Id: I2b71c0989ec88fe98b75d92a3e727719d8879f39