Delete access tokens when demoting users
This commit auto-revokes admin tokens when an admin user is demoted
to a regular user.

It also revokes all access tokens for a given user when that user is
deactivated.

Previously we would let these hang out in bifrost, but this is dangerous
if you deactivate user a@b.c and then subsequently invite a new user
with the email address a@b.c...

AMENDED: changed uri path per review

Change-Id: I68319147544b1c8be91f9a33fccc19c501c7b0a5