Send GetVersReply via transport the request came through
There appear to be weird corner cases where a TCP connection is
broken but one peer does not realize that.

The change that turned GVC/GVR from sync RPC into async request
response had an unexpected side-effect in such situations.
Without the RPC glue, the sender no longer detects that messages
are going into a blackhole and does not trigger pulsing. The
other side meanwhile happily drops packet for the broken connection
because of the datagram interface.

Two clients that would otherwise be perfectly able to talk through
Zephyr find themselves unable to communicate at all because of the
broken TCP connection.

There's more work required to address the weird TCP behavior but
this should get us back to a good enough state in the meantime.

Change-Id: I88fc70bcdcccb927c6b1c28805ebd25605bae137