Fix moving out of a deleted parent
Given path /A/B

device 1 deletes A
device 2 moves B out of A

device 1 tries to readmit B at the old path (i.e. under the trash
folder) and obviously fails.

In the past device 1 would just fail forever and ever to apply
the move. However, with the introduction of NROs, failure to
re-create the deleted hierarchy would result in the top of the
hierarchy being treated as non-representable.

Change-Id: I8cfe1a8fe15494aaed17e93f67bacae2807fd5f1