Disable AVX when building OpenSSL
A bunch of users have hardware running OSX that does not support the AVX
instruction set.  This caused the daemon to get a SIGILL on affected
hardware in the libjingle initialization when RAND_poll used the vxorps
instruction on hardware that did not support it.

The JVM would catch SIGILL, produce a crash dump, and raise a SIGABRT,
which would be detected by the GUI as "exit code 134" (and no additional
information).

-mno-avx is sufficient to convince clang not to use AVX instructions
like vxorps which will cause issues on hardware that we still support.

Change-Id: Ie2c1757e2bd70c2fd38a535e2e07bdcedddf652f