MVP cmd-line tool to register OAuth tokens
This is a minimal python tool for a customer that needs to generate
OAuth access tokens for users, given that users' id and password.

Eventually we could imagine using this for testing purposes, but that
will require pulling in the protobuf dependency in an intelligent way.
This script; it Does Not Do That. Not even a bit.

Change-Id: I7e760b91579686a4260b6cd0b12be30139581a42