Fix no sync and related spurious deletion on Windows
SUPPORT-1439

Failure to close prefix input stream when hashing would cause
move to root anchor to fail. NRO would kick in but would also
fail most of the time. Until it did not (go figure... this is
Windows).

In its infinite wisdom, Windows emits file notifications that
say the file that could not be moved was deleted. Failure to
filter such motifications for NROs would cause the object to
be incorrectly deleted and the deletion would propagate to
other peers.

Fix both the root and proximal cause.
Add test to avoid regression in TimeoutDeletionBuffer.

Change-Id: I4adcab308dd2ffcd76c0e80d1ff30768f69eb053