web: defend against BREACH attack
An attacker with the ability to:

1) Inject partial chosen plaintext into a victim's requests, and
2) Measure the size of encrypted traffic

can leverage information leaked through compression to recover other
parts of the plaintext, provided:

1) The page is served with HTTP-level compression
2) The page reflects some part of the user input in HTTP response bodies
3) The page body contains some secret (such as a CSRF token).

In conjunction with XSS, an attacker could inject plaintext into all
requests, which means that we'd basically have to disable HTTP
compression.  Without XSS, this attack can only be performed on pages
that reflect user input on GET requests.  This would be limited to pages
that change behavior based on URL parameters.

We reflect the `next` URL parameter in the body of /login.  With careful
work, this could be leveraged to reveal the CSRF token of a user through
many requests to /login

As a bonus, this commit cleans up the mess that is the sign in form.
1) Check HTTP method rather than looking for presence of a form value
   to determine if the login form was submitted
2) Remove hidden "next" parameter from the form body, preferring to keep
   it in the URL.
3) Forms with action "" will post to the original URL they were loaded
   from, which will preserve the `next` URL param.

As an addon bonus, the form.submitted nonsense has been removed in its
entirety.

Change-Id: I2a598d1cc67fec5eeea6db6614e83ac24b30fa17