Hardcode default monitoring username
It was folly to set an external.properties field for something that is
not field-serviceable; apparently I missed a localprod case.

So let's just not do that. This doesn't negatively impact our ability
to change the value in the future, at that time we can define a property
and use it.

AMENDED: changed page title to Service status monitor

Change-Id: I42d96e9a7b8de3ed23ca95582da6cf4250c4a1fc