web/bunker: extract deployment folders to config
Part of the ongoing effort to make applications well-behaved and declare
all couplings explicitly in configuration.

This patchset moves the following into values in the *.ini files:

* /opt/repackaging/installers/original, used to determine version
* /opt/bunker/state, a folder used by bunker for storing some state
* /var/aerofs/, a folder containing a couple of flag files pertaining to
  initialization and configuation set by bootstrap and also read by nginx

The first value is of interest to PRIVATE web and bunker instances.
The second and third values are only of interest to bunker.

Change-Id: Ieeb32771dd6b06095fb4176cb9dc905a8ef17b1b