Fix deadlock at startup
Core queue is bounded (1024 entries)
Event processing is only started after all subsystems are
themselves started.
Blocking enqueue into the core queue in a subsystem start()
method must therefore be avoided at all costs.

This issue was noticed in TS scalability tests. It manifests
itself as a hanging daemon. The UI fails to heartbeat and
assumes the dameon is dead so attempts to restart it, which
fails because the hanging daemon is still holding a lock on
the core db.

Change-Id: I9f56300d0a760ed4a2d22d127c92faa6c33c7a52