Prefer VERIFIED channels over CONNECTING ones
If there is no indication of a specific channel to use for an outgoing
packet, we should choose one that is verified instead of one that isn't.

If there are only CONNECTING, that's fine, just choose one.

Start of route-cost structure is included here to keep my life a little simpler.

AMENDED: updated internals from review feedback

Change-Id: I7596db894d61f77f407442589c1ba53a6333572f