Added instruction to turn on API access to mobile install page
Based on https://aerofs.atlassian.net/browse/ENG-1127

If the user is not part of a private deployment, they'll see a
first step telling them to turn on API access for the mobile app
to work, with a link to a blog post with instructions how.

This change also puts the instructions into an ordered list
rather than doing the numbering manually.

Change-Id: I51bdf6f7130cf674646f91ac1bc6f30811e0e0ec