rest-api: extract IUserAuthToken from AuthToken
An AuthToken need only answer the question "does this token have
sufficient permission to authorize this action?" That role has been
extracted to the interface IAuthToken.

However, most things in our system cannot use the interface IAuthToken
directly.  Based on how we've written some of our resources, a number of
our methods require not just knowing hasPermission, but also which user
is making the request, so that we give context-specific responses.  For
that situation, you want your IAuthToken to also return a UserID, as
well as an issuer and unique identifier for the token.  This interface
is IUserAuthToken.

The concrete implementation of these interfaces is renamed to
OAuthToken, as that is the specific implementation of IUserAuthToken
that will get instantiated.  Some resources in the daemon currently
require even more information than IUserAuthToken provides.  For now,
they can continue to expect the injection of an OAuthToken.

Change-Id: I878f8e6ccc59c76d9c75867a3b8c2a6265e92457