User cannot see 'null' in place of a folder name in shared folder UI
Now if the folder's name is given to the message generator as null, it'll show
a more generic message: "I'd like to share a folder with you."

Not sure why the Team Server would show null when the client doesn't--there's
only one mention of the relevant strings in the code base! I don't yet know
how to view the TS locally, and my Java skills are not the best, and it's
possible the underlying bug has been fixed between now and December when the
issue was reported. So long as there's no deep underlying flaw breaking
other things, this fix will fix things from the user's perspective.

Change-Id: I8a65553b6cbfdaf8d39b045037a849aefb4aa4ab