OSUtil: fix Linux architecture detection
Unlike almost every other heuristic we've tried in the past, we can
pretty much rely on the executable's ELF header being valid and an
accurate representation of what architecture we're running on.

This conveniently brings Java's architecture detection into line with
the one that's used by the copier/launcher scripts, which have been
using trustworthy system invariants for quite some time now, with nary a
problem.

Change-Id: I22f1ac4154324bcc63007931b8be9625b06919fd