Giant Bootstrap 3 upgrade of doom
This commit switches web, bunker, and lizard to using Bootstrap 3 instead
of Bootstrap 2. It touches almost every template and frontend file in
the codebase. Yikes.

The biggest changes between 2 and 3 are the grid and the way forms are
styled. Icons, inline elements, and some other things changed as well.

The benefit of this change is twofold:
1.) Since the new splash website uses 3, the rest of our web properties
will now be able to share styles and components with it easily.
2.) 3 is superior to 2 in a number of ways, most importantly in being
a lot more powerful for making responsive layouts and components.

This commit also includes some opportunistic moving of inline styles
into stylesheets, etc etc.

Change-Id: I1adbf20a79f4f1acaf095b856f52c6bf93d24137