Brief attempt to make invitations menu item less confusing
Right now, people don't know whether "My Invitations" refers to
folders they're invited to, people they've invited to share their
folders, or both. Changed the text to "Pending Invitations"; we
can see if that is less confusing for users.

(The latter information is currently not available via the web, and
likely cannot be made available without either insane effort or
using a real API instead of sp to give the website data.)

Later I would like to hide the invitations link in the menu unless
the user has a pending invite, since most of the time it's a
worthless link going to an empty page. However, this is also
currently practically impossible with the current web architecture
for the same reason as above.

Change-Id: I8c73987a3210d004c25676a955c335f9a7a9ed81