Improved desktop client uploading to Dryad.
There were couple problems with previous approach:
- using multiple connections to upload data sequentially is expensive
because each connection needs to perform a SSL handshake.
- compressing to file before streaming is not ideal because it leaves
a footprint and consumes RAM if we somehow cannot delete the log
immediately.
- compressing large files causes 100% cpu usage because Java's zip library
does no thread management (rightly so).

The previous approach was chosen because it was the only approach that I
understood and have verified to work at the time.

I've found another approach that works since, and so I'm updating the code
to use _the_new_way_: using a single connection with chunked transfer
encoding.

The new approach has couple benefits over the old approach:
- single connection means a single SSL handshake reducing connection
overhead.
- using chunked transfer encoding means we don't need to know the file size
before hand, so we can directly stream the compressed output to network.
- streaming compressed output to network applies back-pressure to the
compression layer. Consequently, the cpu usage is reduced.
- nginx automatically piece the pieces before presenting a single stream
to the app layer, so it is transparent and simple to work with.

Change-Id: Ia940a534e20f6f763056a1b9cf1500cf7814f15e