Proactively connect unicast channels
Add a device presence listener that, upon notification of a device
online on multicast, tries to create a client connection to that device.

In general, this change has no significant user-visible impact. This is
preliminary work for simplifying our view of presence.

AMENDED: use a channel future throughout the ChannelDirectory/Unicast path
AMENDED: added a missing test class

Change-Id: Idb35fb425c7fa5095b63c70486ed8966470d2b9b