New icon for shared folders; Yosemite shared/root folder graphics exist
We now have a special sharing icon instead of using a wifi icon for shared folders.

Also, towards the goal of standardizing on the non-gradient version of the aerofs
logo, everything in resource/client/all/icons/ (as well as the client specific folder
graphics) uses the flat design aerofs logo. Still need to update things like the
installation splash drive graphic or the tutorial, but, a good start.

Change-Id: I66e1a5eb15161f3ceb993b2b41ed3935a1f95480